Split file index tests from production module
